fixing scripts
