--- console/run.ts
@@ -1,16 +1,0 @@
-const decoder = new TextDecoder();
-
-export async function run(command: string[]) {
-  const p = Deno.run({
-    cmd: command,
-    stderr: "piped",
-  });
-
-  const status = await p.status()
-
-  if (status.code === 1) {
-    console.log(decoder.decode(await p.stderrOutput()));
-  }
-
-  p.close();
-}
